updates to database upload
Updated bitstream.py to have a callable function that can be used by other scripts for extracting bitstream metrics which returns the metrics as a tuple

Updated model_predict.py to allow for users to specify metrics that may be provided to the model, but should NOT be used in making predictions. These will only be passed through to the database.
Added code for updating the database's logstring in model_predict

Updated quality_extractor_mp4s.py to also calculate bitrate, framerate, and resolution. Note that this is calculated on samples from a livestream rather than calculated directly from the livestream